feat: stabilize version switcher paths
